+++ src/services/__mocks__/fulfillment/index.js
@@ -1,3 +1,4 @@
 export { fulfillmentMock } from "./mock"
+export { fulfillmentRepoMock } from "./repo"
 export { fulfillmentServiceMock } from "./service"
 export { fulfillmentSchema, fulfillmentItemSchema } from "./schema"
