--- src/services/__mocks__/totals-service.js
@@ -1,5 +1,0 @@
-import { lineItemTotalsTemplate } from "./templates"
-
-export const totalsServiceMock = () => ({
-  getLineItemTotals: jest.fn(async () => lineItemTotalsTemplate()),
-})
