+++ src/services/__mocks__/totals/index.js
@@ -1,0 +1,2 @@
+export { totalsServiceMock } from "./service"
+export { lineItemTotalsSchema } from "./schema"
