+++ src/services/__mocks__/totals/service.js
@@ -1,0 +1,5 @@
+import { lineItemTotalsSchema } from "./schema"
+
+export const totalsServiceMock = () => ({
+  getLineItemTotals: jest.fn(async () => lineItemTotalsSchema()),
+})
